MLBackend/requirements.txt
scikit-learn
nltk
pandas
Flask
requests
scipy
python-dotenv